feat: improve UX
